Introduce common StepList for Step bundles and Workflows
